Event Broadcasters
Added a file where I keep common errors I run into for debugging
reference.

Event broadcasters are a new object that exposes when
certain events happen that you might want to track.  Current ones
are crew death animation start and crew death animation end.

Working on a blood thing, and a menu to enable it in.  Or I could add it
to the toggle menu I guess.